+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Planned]
 
+- Custom global allocator.
+- Environment variables.
+
+## [0.1.0-alpha.6] - 2025-03-09
+
+### Added
+
+- Heap-based allocation and data types.
+- `mash` process execution.
+
+### Changed
+
+- Tests now return with success or failure.
+
 ## [0.1.0-alpha.5] - 2025-03-08
 
 ### Added
@@ -15,10 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show basename in `mash` prompt.
 - `poweroff` command.
 - `reboot` command.
-
-### Changed
-
-### Removed
 
 ## [0.1.0-alpha.4] - 2025-03-07